ml_model/requirements.txt
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.0
requests>=2.31.0
python-dotenv>=1.0.0
setuptools>=69.0.0 